Refactored spring java config to work the same way as the xml version.
