+++ SpringIoC/pom.xml
@@ -25,7 +25,7 @@
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context</artifactId>
-            <version>4.2.5.RELEASE</version>
+            <version>4.3.5.RELEASE</version>
         </dependency>
     </dependencies>
