+++ SpringIoC/src/main/java/kvhadzhiev/ioc/spring/javaconfig/util/FileLogger.java
@@ -1,5 +1,7 @@
 package kvhadzhiev.ioc.spring.javaconfig.util;
 
+import org.springframework.beans.factory.annotation.Autowired;
+
 // start second logger(use prototype scope for logger)
 public class FileLogger {
 
@@ -7,10 +9,7 @@ public class FileLogger {
 
     }
 
-    public FileLogger(ActionLogger actionLogger) {
-        actionLogger.init();
-    }
-
+    @Autowired
     public void setLogger(ActionLogger actionLogger) {
         actionLogger.init();
     }
